Update extension: Quick Access
