+++ extensions/quick-access/src/components/quick-access-empty-view.tsx
@@ -1,4 +1,4 @@
-import { Action, ActionPanel, Icon, List } from "@raycast/api";
+import { Action, ActionPanel, Grid, Icon, List } from "@raycast/api";
 import React, { Dispatch, SetStateAction } from "react";
 import { pinFiles } from "../pin";
 import { refreshNumber } from "../hooks/hooks";
@@ -6,13 +6,33 @@ import { ActionRemoveAllDirectories } from "./action-on-files";
 import { DirectoryWithFileInfo } from "../types/types";
 
 export function QuickAccessEmptyView(props: {
+  layout: string;
   title: string;
   description: string;
   setRefresh: Dispatch<SetStateAction<number>>;
   directoryWithFiles: DirectoryWithFileInfo[];
 }) {
-  const { title, description, setRefresh, directoryWithFiles } = props;
-  return (
+  const { layout, title, description, setRefresh, directoryWithFiles } = props;
+  return layout === "Grid" ? (
+    <Grid.EmptyView
+      icon={{ source: { light: "empty-view-icon.svg", dark: "empty-view-icon@dark.svg" } }}
+      title={title}
+      description={description}
+      actions={
+        <ActionPanel>
+          <Action
+            title={"Pin"}
+            icon={Icon.Pin}
+            onAction={async () => {
+              await pinFiles([], false);
+              setRefresh(refreshNumber());
+            }}
+          />
+          {directoryWithFiles.length != 0 && <ActionRemoveAllDirectories setRefresh={setRefresh} />}
+        </ActionPanel>
+      }
+    />
+  ) : (
     <List.EmptyView
       icon={{ source: { light: "empty-view-icon.svg", dark: "empty-view-icon@dark.svg" } }}
       title={title}
@@ -23,7 +43,7 @@ export function QuickAccessEmptyView(props: {
             title={"Pin"}
             icon={Icon.Pin}
             onAction={async () => {
-              await pinFiles(false);
+              await pinFiles([], false);
               setRefresh(refreshNumber());
             }}
           />
