+++ extensions/quick-access/src/types/preferences.ts
@@ -1,4 +1,7 @@
 export interface Preferences {
+  layout: string;
+  columns: number;
+  itemInset: string;
   showOpenFolders: boolean;
   autoCopyLatestFile: boolean;
   rememberTag: boolean;
